+++ sdk/identity/identity/src/msal/flows.ts
@@ -18,6 +18,7 @@ export interface MsalFlowOptions {
   authorityHost?: string;
   authenticationRecord?: AuthenticationRecord;
   disableAutomaticAuthentication?: boolean;
+  disableAuthorityValidation?: boolean;
 }
 
 /**