+++ sdk/keyvault/keyvault-admin/src/lro/selectiveKeyRestore/operation.ts
@@ -10,10 +10,10 @@ import {
   KeyVaultSelectiveKeyRestoreResult,
 } from "../../backupClientModels";
 import {
-  KeyVaultClientRestoreStatusResponse,
-  KeyVaultClientSelectiveKeyRestoreOperationOptionalParams,
-  KeyVaultClientSelectiveKeyRestoreOperationResponse,
   RestoreOperation,
+  RestoreStatusResponse,
+  SelectiveKeyRestoreOperationOptionalParams,
+  SelectiveKeyRestoreOperationResponse,
 } from "../../generated/models";
 import { AbortSignalLike } from "@azure/abort-controller";
 import { KeyVaultClient } from "../../generated/keyVaultClient";
@@ -74,8 +74,8 @@ export class KeyVaultSelectiveKeyRestorePollOperation extends KeyVaultAdminPollO
    */
   private selectiveRestore(
     keyName: string,
-    options: KeyVaultClientSelectiveKeyRestoreOperationOptionalParams
-  ): Promise<KeyVaultClientSelectiveKeyRestoreOperationResponse> {
+    options: SelectiveKeyRestoreOperationOptionalParams
+  ): Promise<SelectiveKeyRestoreOperationResponse> {
     return withTrace("selectiveRestore", options, (updatedOptions) =>
       this.client.selectiveKeyRestoreOperation(this.vaultUrl, keyName, updatedOptions)
     );
@@ -84,10 +84,7 @@ export class KeyVaultSelectiveKeyRestorePollOperation extends KeyVaultAdminPollO
   /**
    * Tracing the restoreStatus operation.
    */
-  private restoreStatus(
-    jobId: string,
-    options: OperationOptions
-  ): Promise<KeyVaultClientRestoreStatusResponse> {
+  private restoreStatus(jobId: string, options: OperationOptions): Promise<RestoreStatusResponse> {
     return withTrace("restoreStatus", options, (updatedOptions) =>
       this.client.restoreStatus(this.vaultUrl, jobId, updatedOptions)
     );