+++ sdk/keyvault/keyvault-certificates/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 4.4.0-beta.3 (Unreleased)
+## 4.4.0 (Unreleased)
 
 ### Features Added
 
@@ -10,6 +10,8 @@
 
 ### Other Changes
 
+- This release updates `CertificateClient` to support service version 7.3 by default.
+
 ## 4.4.0-beta.2 (2021-11-09)
 
 ### Features Added
@@ -18,7 +20,7 @@
 
 ### Other Changes
 
-- Updated CertificateClient to use service version 7.3-preview by default.
+- Updated CertificateClient to use service version 7.3 by default.
 
 ## 4.4.0-beta.1 (2021-09-07)